#11 style: SideBar을 LeftBar, RightBar로 분리 / LeftBar의 state를 DefaultMode, SearchMode로 분리
