+++ src/components/organisms/SideBar/index.tsx
@@ -1,20 +1,11 @@
-import { useState, useRef, useEffect } from 'react';
-import SearchBar from 'src/components/organisms/SearchBar';
-import SearchResult from 'src/components/organisms/SearchResult';
-import Dropdown from 'src/components/organisms/Dropdown';
+import { useState, useRef } from 'react';
 import SideBarItem from 'src/components/molecules/SideBarItem';
-import IconButton from 'src/components/molecules/IconButton';
+import Dropdown from 'src/components/organisms/Dropdown';
 import SideBarDivider from 'src/components/molecules/SideBarDivider';
-import List from 'src/components/molecules/List';
-import ListItem from 'src/components/atoms/ListItem';
-import ToggleButton from 'src/components/atoms/ToggleButton';
-import useNodesData from 'src/hooks/useNodeData';
 import { IconType } from 'src/types/icon';
+import SearchMode from './Searchmode';
+import DefaultMode from './DefaultMode';
 import * as s from './style';
-
-type SideBarModeProps = {
-  changeMode: () => void;
-}
 
 const exampleToggleButtons: IconType[] =
   Array.from({ length: 10 }).map(() => 'tree');
@@ -22,80 +13,27 @@ const exampleToggleButtons: IconType[] =
 const dropdownItems: string[] =
   Array.from({ length: 10 }).map((_, idx) => `item ${idx}`);
 
-const exampleListItems: JSX.Element[] =
-  Array.from({ length: 10 }).map(() => <ListItem itemTitle='List item' />);
-
-
-const SearchMode = ({ changeMode }: SideBarModeProps) => {
-  const { nodes } = useNodesData();
-  const [searchInput, setSearchInput] = useState<string>('');
-  const [searchResults, setSearchResults] = useState<string[]>([]);
-
-  useEffect(() => {
-    if (searchInput.length) {
-      setSearchResults(
-        [...Object.keys(nodes)].filter(v => v.includes(searchInput)),
-      );
-    } else {
-      setSearchResults([]);
-    }
-  }, [nodes, searchInput]);
-
-
-  const handleDropdownOptionClick = (optionValue: string) => {
-    alert(optionValue);
-  };
-
-  const collapseSearchBar = () => {
-    changeMode();
-    setSearchInput('');
-    setSearchResults([]);
-  };
-
-  return (
-    <div>
-      <SideBarItem title='' element={
-        <SearchBar
-          searchInput={searchInput}
-          setSearchInput={setSearchInput}
-          collapseSearchBar={collapseSearchBar}
-        />
-      } />
-      <SideBarDivider />
-      <SearchResult results={searchResults} handleOptionClick={handleDropdownOptionClick} />
-
-    </div>
-  );
-};
-
-
-const DefaultMode = ({ changeMode }: SideBarModeProps) => (
-  <>
+const RightBar = () => (
+  <s.SideBar>
     <SideBarItem
-      title='Search'
-      element={<IconButton icon='search' onClick={changeMode} />} />
+      title='Map Theme'
+    />
     <SideBarDivider />
     <SideBarItem
       title='Theme direction'
       element={<Dropdown icons={exampleToggleButtons} items={dropdownItems} />}
     />
-    <SideBarDivider />
     <SideBarItem
       title='Theme colors'
-      element={<ToggleButton onClick={() => alert('Theme colors')} />}
     />
-    <SideBarDivider />
     <SideBarItem
-      title='Pages'
-      element={<List items={exampleListItems} />}
-      noPadding
+      title='Image'
     />
-    <SideBarDivider />
-  </>
+  </s.SideBar>
 );
 
 
-const SideBar = () => {
+const LeftBar = () => {
   const [searchMode, setSearchMode] = useState<boolean>(false);
   const sideBarRef = useRef<HTMLDivElement>(null);
   const changeMode = () => setSearchMode(!searchMode);
@@ -110,4 +48,4 @@ const SideBar = () => {
   );
 };
 
-export default SideBar;
+export { RightBar, LeftBar };
